requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
lxml>=5.1.0
python-dateutil>=2.8.2
openai>=1.0.0
matplotlib>=3.7.0
